+++ CHANGELOG.md
@@ -39,6 +39,8 @@
   - [#1641](https://github.com/wix-incubator/rich-content/pull/1641) Gallery Plugin Docs
 - `divider`
   - [#1634](https://github.com/wix-incubator/rich-content/pull/1634) Divider Plugin Docs
+- `file-upload`
+  - [#1644](https://github.com/wix-incubator/rich-content/pull/1644) File Upload Plugin Docs
 
 ## 8.0.1 (Oct 14, 2020)
 ### :rocket: New Feature